Remove service triggers
Since we're currently not using procd to start and monitor the service, the service triggers are not necessary at the moment. In fact, it appears that for a few users, these triggers caused qosmate to be stopped by procd during boot. To prevent this, the service triggers have been temporarily removed. Basically, everything should be handled through the hotplug script and the firewall include.